+++ hw/xwin/winmultiwindowwndproc.c
@@ -1229,7 +1229,7 @@ winUpdateWindowPosition(HWND hWnd, HWND * zstyle)
     SetRect(&rcNew, iX, iY, iX + iWidth, iY + iHeight);
 
     winDebug("winUpdateWindowPosition - drawable extent (%d, %d)-(%d, %d)\n",
-             rcNew.left, rcNew.top, rcNew.right, rcNew.bottom);
+             (int)rcNew.left, (int)rcNew.top, (int)rcNew.right, (int)rcNew.bottom);
 
     AdjustWindowRectEx(&rcNew, GetWindowLongPtr(hWnd, GWL_STYLE), FALSE,
                        GetWindowLongPtr(hWnd, GWL_EXSTYLE));
@@ -1248,7 +1248,7 @@ winUpdateWindowPosition(HWND hWnd, HWND * zstyle)
     }
 
     winDebug("winUpdateWindowPosition - Window extent (%d, %d)-(%d, %d)\n",
-             rcNew.left, rcNew.top, rcNew.right, rcNew.bottom);
+             (int)rcNew.left, (int)rcNew.top, (int)rcNew.right, (int)rcNew.bottom);
 
     /* Position the Windows window */
     SetWindowPos(hWnd, *zstyle, rcNew.left, rcNew.top,
